refactor(language): rename stored-language module to language-store
Aligns the module name with the *-store sibling convention
(speech-store, playback-store). The Stored function names stay,
since they disambiguate raw-store access from the context API.

Also drops a redundant explanatory comment in the board-set-index
loader test; the assertion guards itself.
